+++ ProjetoBlazor/Models/PedidoItem.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Numerics;
 
 namespace ProjetoBlazor.Models
 {
@@ -6,7 +7,7 @@ namespace ProjetoBlazor.Models
     {
         [Key]
         public int Controle { get; set; }
-        public int ControlePedido { get; set; }
+        public Int64 ControlePedido { get; set; }
         public int Item { get; set; }
         public int ProdutoCodigo { get; set; }
         public string Descricao { get; set; } = "";
